Add EventClient.trackUserExpired function
Need to be able to emit the event from teachers once the trial-expired
page has been rendered.
Given that this library does not yet set up the fixtures required for
testing various scenarios, the real client does not assert that the
correct request has been made.
This is a simple POST without any body, and without a response so I
think it's OK to skip this for the moment.

[#169737339]