Fix how FakeCartsClient uses an empty cart
Use the CartsFactory to get a new empty cart when needed.
The previous solution had a global `emptyCart` object which we copied when
creating a new cart. Because of how assignment works it meant that the
contained `items` array was referenced in the new cart, and when we
pushed new items to it, the original `emptyCart.items` was also updated.
This resulted that when in future tests we created a new fake client the
empty cart already contained some items.

[#175616338]